requirements.txt
streamlit
supabase
openai
python-dotenv
pandas
numpy
plotly
httpx
postgrest